+++ resources/views/admin/ticket/index.blade.php
@@ -28,7 +28,6 @@
                             :checked="!empty($showInactive)"
                             :noWrapper="true"
                             :inline="true"
-                            inputClass="h-4 w-4 rounded mt-0"
                             labelClass="text-sm pt-0"
                             onchange="this.form.submit()"
                         />
@@ -39,7 +38,6 @@
                             :checked="!empty($groupByWorkshop)"
                             :noWrapper="true"
                             :inline="true"
-                            inputClass="h-4 w-4 rounded mt-0"
                             labelClass="text-sm pt-0"
                             onchange="this.form.submit()"
                         />