+++ resources/views/admin/user/index.blade.php
@@ -15,7 +15,6 @@
                         :checked="!empty($showGhostUsers)"
                         :noWrapper="true"
                         :inline="true"
-                        inputClass="h-4 w-4 rounded mt-0"
                         labelClass="text-sm pt-0"
                         onchange="this.form.submit()" />
                     <div class="flex flex-col sm:flex-row sm:items-center gap-2">